Rename file for naming convention & remove unnecessary tests
